@@ -1,16 +1,25 @@
 {
-  "name": "my-portfolio-works-new",
+  "name": "my-portfolio-works",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/classnames": "^2.2.3",
-    "@types/react-transition-group": "^2.0.9",
-    "node-sass-chokidar": "^0.0.3",
-    "npm-run-all": "^4.1.2",
-    "react": "^16.2.0",
+    "animate.css": "^3.6.1",
+    "classnames": "^2.2.5",
+    "font-awesome": "^4.7.0",
+    "github-markdown-css": "^2.10.0",
+    "react": "^16.3.2",
+    "react-animated-css": "^1.0.4",
     "react-anime": "^2.1.0",
-    "react-dom": "^16.2.0",
-    "react-scripts-ts": "2.13.0"
+    "react-canvas": "^1.3.0",
+    "react-dom": "^16.3.2",
+    "react-id-swiper": "^1.6.2",
+    "react-markdown": "^3.3.0",
+    "react-redux": "^5.0.7",
+    "react-scripts-ts": "2.15.1",
+    "react-show": "^2.0.4",
+    "react-transition-group": "^2.3.1",
+    "redux": "^4.0.0",
+    "sanitize.css": "^5.0.0"
   },
   "scripts": {
     "start": "npm-run-all -p watch-css start-js",
@@ -40,34 +49,20 @@
     "*.json": ["prettier --write", "git add"]
   },
   "devDependencies": {
-    "@types/jest": "^22.1.2",
-    "@types/node": "^9.4.5",
-    "@types/react": "^16.0.36",
-    "@types/react-dom": "^16.0.3",
-    "@types/react-redux": "^5.0.14",
-    "animate.css": "^3.6.1",
-    "classnames": "^2.2.5",
-    "eslint": "^4.17.0",
-    "font-awesome": "^4.7.0",
-    "github-markdown-css": "^2.10.0",
+    "@types/classnames": "^2.2.3",
+    "@types/jest": "^22.2.3",
+    "@types/node": "^10.0.3",
+    "@types/react": "^16.3.13",
+    "@types/react-dom": "^16.0.5",
+    "@types/react-redux": "^5.0.19",
+    "@types/react-transition-group": "^2.0.9",
+    "eslint": "^4.19.1",
     "husky": "^0.14.3",
-    "lint-staged": "^6.1.0",
-    "prettier": "^1.10.2",
-    "react-animated-css": "^1.0.4",
-    "react-canvas": "^1.3.0",
-    "react-id-swiper": "^1.6.1",
-    "react-markdown": "^3.3.0",
-    "react-redux": "^5.0.6",
-    "react-show": "^2.0.4",
-    "react-transition-group": "^2.3.0",
-    "redux": "^3.7.2",
-    "sanitize.css": "^5.0.0",
-    "tslint-config-prettier": "^1.7.0",
-    "typescript": "^2.7.1"
-  },
-  "engines": {
-    "node": "9.4.0",
-    "npm": "*",
-    "yarn": "*"
+    "lint-staged": "^7.0.5",
+    "node-sass-chokidar": "^1.2.2",
+    "npm-run-all": "^4.1.2",
+    "prettier": "^1.12.1",
+    "tslint-config-prettier": "^1.12.0",
+    "typescript": "^2.8.3"
   }
 }
