@@ -2,7 +2,6 @@ import * as React from 'react';
 import * as Swiper from 'react-id-swiper';
 
 export default class Header extends React.Component {
-  /* tslint:disable no-any */
   private swiper: any;
 
   constructor(props: {}) {
@@ -11,12 +10,12 @@ export default class Header extends React.Component {
     this.goPrev = this.goPrev.bind(this);
   }
 
-  render(): JSX.Element {
+  public render(): JSX.Element {
     const params = {
-      slidesPerView: 'auto',
-      spaceBetween: 10,
       freeMode: true,
       loop: true,
+      slidesPerView: 'auto',
+      spaceBetween: 10,
     };
 
     const languages: string[] = [
@@ -52,7 +51,9 @@ export default class Header extends React.Component {
             <Swiper
               {...params}
               ref={(node: any) => {
-                if (node) this.swiper = node.swiper;
+                if (node) {
+                  this.swiper = node.swiper;
+                }
               }}
             >
               {rows}
